Move HTML rendering into the server
Not sure why this is polluting the domain and common agent/server code.
