+++ server/src/main/java/com/thoughtworks/go/server/presentation/html/Htmlable.java
