+++ server/src/main/java/com/thoughtworks/go/server/presentation/models/JobDetailPresentationModel.java
@@ -19,8 +19,8 @@ import com.thoughtworks.go.config.Tabs;
 import com.thoughtworks.go.config.TrackingTool;
 import com.thoughtworks.go.domain.*;
 import com.thoughtworks.go.domain.exception.IllegalArtifactLocationException;
+import com.thoughtworks.go.server.presentation.html.HtmlRenderer;
 import com.thoughtworks.go.server.service.ArtifactsService;
-import com.thoughtworks.go.util.DirectoryReader;
 import com.thoughtworks.go.util.TimeConverter;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
