+++ server/src/test-fast/java/com/thoughtworks/go/domain/FileDirectoryEntryTest.java
@@ -16,10 +16,11 @@
 
 package com.thoughtworks.go.domain;
 
-import com.thoughtworks.go.server.presentation.models.HtmlRenderer;
+import com.thoughtworks.go.server.presentation.html.HtmlRenderer;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class FileDirectoryEntryTest {
     @Test
